backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
PyJWT==2.8.0
Werkzeug==2.3.7
pymongo==4.5.0
pyserial==3.5
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2